Add tests for Handlers (WIP).
Refactor:
- extract appService and mkAppEnv to App
- extract newtype App to App.Types
